refactor resolvesignable

+++ Assets/Plugins/Flow/FCL/Models/Authz/Account.cs
@@ -12,10 +12,10 @@ namespace Flow.FCL.Models.Authz
         public string TempId { get; set; }
         
         [JsonProperty("keyId")]
-        public int KeyId { get; set; }
+        public uint KeyId { get; set; }
         
         [JsonProperty("sequenceNum")]
-        public int SequenceNum { get; set; }
+        public ulong SequenceNum { get; set; }
         
         [JsonProperty("signature")]
         public string Signature { get; set; }
